Restaurant

+++ src/main/java/com/molvenolakeresort/models/restaurant/order/Order.java
@@ -15,27 +15,34 @@ public class Order {
     @Id @GeneratedValue
     private Long idOrder;
     private SupplierOrderStatus status;
-    private int totalPrice;
+    private double totalPrice;
 
     @ManyToOne
     private Room room;
+
     @OneToOne
     private RestaurantTable restaurantTable;
 
-    //EVENT????
-
     @OneToMany(
-        mappedBy = "order",
-        cascade = CascadeType.ALL,
-        orphanRemoval = true
+            mappedBy = "order",
+            cascade = CascadeType.ALL,
+            orphanRemoval = true
     )
+
     private List<MenuItemOrder> menuItem_orders;
 
     public Long getIdOrder() {
+
         return idOrder;
+
     }
 
+
+
     public SupplierOrderStatus getStatus() {
+
         return status;
+
     }
+
 }
